requirements.txt
pdfplumber~=0.11.1
pandas~=2.2.2
tk
PyQt5~=5.15.10
setuptools~=68.2.0
camelot-py[cv]
PyQt5
PyPDF2<3.0.0
opencv-python
PyMuPDF
pytesseract~=0.3.10
openai~=0.27.0
Pillow~=10.3.0